add dropdown options to recipes store
