+++ src/Store/RecipesStore/types.ts
@@ -30,3 +30,7 @@ export type Option = {
     key: string;
     value: string;
 };
+
+export type DropdownCounts = {
+    [key: string]: number
+}
